quiet compiler warnings

+++ src/detection/uptime/uptime_qnx.c
@@ -8,7 +8,7 @@ const char* ffDetectUptime(FFUptimeResult* result) {
     if (clock_gettime(CLOCK_MONOTONIC, &tp) == -1)
         return nullptr;
 
-    result->uptime = (uint64_t)tp.tv_sec * 1000 + (uint64_t)tp.tv_nsec / 1e6;
+    result->uptime = (uint64_t)tp.tv_sec * 1000 + (uint64_t)((double) tp.tv_nsec / 1e6);
     result->bootTime = ffTimeGetNow() - result->uptime;
     return nullptr;
 }
